+++ src/py_gql/_pkg.py
@@ -6,7 +6,7 @@ Package information.
 __title__ = "py_gql"
 __description__ = "Comprehensive GraphQL implementation for Python."
 __url__ = "https://github.com/lirsacc/py-gql"
-__version__ = "0.6.0"
+__version__ = "0.6.1"
 __author__ = "Charles Lirsac"
 __author_email__ = "c.lirsac@gmail.com"
 __license__ = "MIT"